Refactored chapter mode with new routing and template; validation utilities.
